Function Calls

SDK Functionality

The Synchrony SDK is split into the following functional groups:

Authentication and authorization calls

User calls

Robot calls

Session calls

Cyphon calls

Authentication and Authorization

Authentication and authorization calls are the first step in connecting to Synchrony. The calls are used to create tokens used for subsequent API calls.

Auth Structures

SynchronySDK.AuthRequest — Type.

The structure used to perform an authentication request.

source

SynchronySDK.AuthResponse — Type.

The structure returned from an authentication call.

source

Auth Functions

SynchronySDK.authenticate — Function.

authenticate(config::SynchronyConfig, authRequest::AuthRequest)::AuthResponse

Authenticates a user and produces a token for further user. Return: The authentication token.

source

SynchronySDK.refreshToken — Function.

refreshToken(config::SynchronyConfig, authResponse::AuthResponse)::AuthResponse

Refreshes a token from an older token. Use this when the token is about to expire. Return: The updated token.

source

User

User calls are used to create, update, retrieve, or delete users from Synchrony. It is also used to retrieve runtime configuration information per user (e.g. streaming connection information).

User Structures

SynchronySDK.UserRequest — Type.

The structure used for user requests.

source

SynchronySDK.UserResponse — Type.

The response structure for user calls.

source

SynchronySDK.UserConfig — Type.

Runtime configuration parameters for connecting to the Synchrony API.

source

SynchronySDK.KafkaConfig — Type.

Runtime configuration parameters for connecting to the streaming API.

source

SynchronySDK.ErrorResponse — Type.

Standardized error response for any request. This is returned if any requests fail server-side.

source

User Functions

SynchronySDK.createUser — Function.

createUser(config::SynchronyConfig, auth::AuthResponse, user::UserRequest)::UserResponse

Create a user in Synchrony. Return: Returns the created user.

source

SynchronySDK.getUser — Function.

getUser(auth::AuthResponse, userId::String)::UserResponse

Gets a user given the user ID. Return: The user for the given user ID.

source

SynchronySDK.updateUser — Function.

updateUser(auth::AuthResponse, user::User)::UserResponse

Update a user. Return: The updated user from the service.

source

SynchronySDK.deleteUser — Function.

deleteUser(auth::AuthResponse, userId::String)::UserResponse

Delete a user given a user ID. Return: The deleted user.

source

SynchronySDK.getUserConfig — Function.

getUserConfig(auth::AuthResponse, userId::String)::UserConfig

Get a user config given a user ID. The user config contains all the runtime parameters for any robot. Return: The user config.

source

Robot

Robot calls are used to create, update, retrieve, or delete robots related to users in Synchrony.

Session

Session calls are used to ingest or retrieve runtime data from Synchrony. Every running robot saves data against a session, and the Session calls allow a user to retrieve data across all sessions.

Cyphon

Cyphon calls allow a user to retrieve more complex data by performing Cypher queries against the saved graph.