Common Structures and Functions

Common Structures and Functions

The following are common structures and functions that are used across all the services.

Structures

GraffSDK.SynchronyConfig — Type.

Configuration structure for Synchrony, it defines the Synchrony Web API endpoint and port.

source

GraffSDK.ErrorResponse — Type.

Standardized error response for any request. This is returned if any requests fail server-side.

source

Functions

GraffSDK.loadConfigFile — Function.

loadConfigFile(filename)

Load a config file from a file name.

source

GraffSDK.getStatus — Function.

getStatus(config)

Get the status of the Synchrony service.

source

GraffSDK.printStatus — Function.

printStatus(config)

Print the current service status.

source