Robot Service

Robot Service

Robot calls are used to create, update, retrieve, or delete robots related to users in the GraffSDK.

Robot Structures

GraffSDK.RobotRequest — Type.

The structure used for robot requests.

source

GraffSDK.RobotResponse — Type.

The structure returned when any robot requests are made.

source

GraffSDK.RobotsResponse — Type.

A list of robots provided by the /robots request.

source

Robot Functions

GraffSDK.getRobots — Function.

getRobots(config)

Gets all robots managed by the specified user. Return: A vector of robots for a given user.

source

GraffSDK.isRobotExisting — Function.

isRobotExisting(config, robotId)

Return: Returns true if the robot exists already.

source

isRobotExisting(config)

Return: Returns true if the robot in the config exists already.

source

GraffSDK.getRobot — Function.

getRobot(config, robotId)

Get a specific robot given a user ID and robot ID. Will retrieve config.robotId by default. Return: The robot for the provided user ID and robot ID.

source

getRobot(config)

Get a specific robot given a user ID and robot ID. Will retrieve config.robotId by default. Return: The robot for the provided user ID and robot ID.

source

GraffSDK.addRobot — Function.

addRobot(config, robot)

Create a robot in Synchrony and associate it with the given user. Return: Returns the created robot.

source

GraffSDK.updateRobot — Function.

updateRobot(config, robot)

Update a robot. Return: The updated robot from the service.

source

GraffSDK.deleteRobot — Function.

deleteRobot(config, robotId)

Delete a robot given a robot ID. Return: The deleted robot.

source

Robot Configuration Functions

GraffSDK.getRobotConfig — Function.

getRobotConfig(config, robotId)

Will retrieve the robot configuration (user settings) for the given robot ID. Return: The robot config for the provided user ID and robot ID.

source

getRobotConfig(config)

Will retrieve the robot configuration (user settings) for the default robot ID. Return: The robot config for the provided user ID and robot ID.

source

GraffSDK.updateRobotConfig — Function.

updateRobotConfig(config, robotId, robotConfig)

Update a robot configuration. Return: The updated robot configuration from the service.

source

updateRobotConfig(config, robotConfig)

Update a robot configuration. Return: The updated robot configuration from the service.

source