User Service

User Service

User calls are used to create, update, retrieve, or delete users from Synchrony. It is also used to retrieve runtime configuration information per user (e.g. streaming connection information).

User Structures

GraffSDK.UserRequest — Type.

The structure used for user requests.

source

GraffSDK.UserResponse — Type.

The response structure for user calls.

source

User Functions

GraffSDK.getUser — Function.

getUser(config, userId)

Gets a user given the user ID. Return: The user for the given user ID.

source

GraffSDK.updateUser — Function.

updateUser(config, user)

Update a user. Return: The updated user from the service.

source

GraffSDK.deleteUser — Function.

deleteUser(config, userId)

Delete a user given a user ID. NOTE: All robots must be deleted first, the call will fail if robots are still associated to the user. Return: The deleted user.

source