Robot Service

Robot Service

Robot calls are used to create, update, retrieve, or delete robots related to users in Synchrony.

Robot Structures

SynchronySDK.RobotRequest — Type.

The structure used for robot requests.

source

SynchronySDK.RobotResponse — Type.

The structure returned when any robot requests are made.

source

SynchronySDK.RobotsResponse — Type.

A list of robots provided by the /robots request.

source

Robot Functions

getRobots
getRobot
createRobot
updateRobot
deleteRobot